Finished splitting rx/tx modules
